# [HUDI-625] Fixing performance issues around DiskBasedMap & kryo