@lamber-ken 2020-02-24
The (4000000 enties) upsert time improve from ~hours to ~5min.