@lamber-ken 2020-02-25
> WOW! @lamber-ken Thanks for investigating this! Your finding make a lot of sense to me. This change definitely gonna boost the upsert performance a lot.

Thanks, we are all having fun here : )