@lamber-ken 2020-02-25
> > > The (4000000 enties) upsert time improve from ~hours to ~5min.
> 
> @lamber-ken this is for each upsert or the entire sequence of 1 insert + 2 upserts? I ask because, I got ~2 minutes on my mac with the custom serializers

just 1 upsert operation, range 2min~5min. I placed the max cost time here.