@lamber-ken 2020-02-25
> classes extending BaseAvroPayload may be okay. but we allow users to implement their own payload classes.. if they don't have data converted as byte[], do we need to register them etc?

It's right way to register the class to kryo, but if the custom payload doesn't have a no-arg constructor, it will not work.

And, we already set the instantiator strategy, kryo built-in serializers have high performance.
```
kryo.setInstantiatorStrategy(new DefaultInstantiatorStrategy(new StdInstantiatorStrategy()));
```